requirements.txt
azure-cognitiveservices-speech==1.28.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
Werkzeug==2.3.4
